extract navigation, autosave, and bp helpers
